+++ test/pdfs/table/colspan.js
@@ -1,5 +1,5 @@
 module.exports = function(doc, { lorem })  {
-  const table = doc.table({
+  let table = doc.table({
     widths: [null, null, null],
     borderWidth: 1,
     borderVerticalWidths: [2, 4, 6, 8],
@@ -41,6 +41,25 @@ module.exports = function(doc, { lorem })  {
     row.cell('First', { fontSize: 20, padding: 10 })
     row.cell('Second', { fontSize: 20, padding: 10 })
     row.cell('Third', { fontSize: 20, padding: 10 })
+  }
+
+  // test issue #99 (colspan without borderVerticalWidths)
+  table = doc.table({
+    widths: [null, null, null],
+    borderWidth: 1,
+  })
+
+  {
+    const row = table.row()
+    row.cell('First', { fontSize: 20, padding: 10 })
+    row.cell('Second', { fontSize: 20, padding: 10 })
+    row.cell('Third', { fontSize: 20, padding: 10 })
+  }
+
+  {
+    const row = table.row()
+    row.cell('First', { fontSize: 20, padding: 10, colspan: 2 })
+    row.cell('Second', { fontSize: 20, padding: 10 })
   }
 }